updated image.js model
