+++ models/Image.js
@@ -1,42 +1,52 @@
-const { Model, DataTypes } = require('sequelize');
-const sequelize = require('../config/connection');
+const { Model, DataTypes } = require("sequelize");
+const sequelize = require("../config/connection");
 
 class Image extends Model {}
 
 Image.init(
-  {
-    id: {
-      type: DataTypes.INTEGER,
+	{
+		id: {
+			type: DataTypes.INTEGER,
+			allowNull: false,
+			primaryKey: true,
+			autoIncrement: true,
+		},
+		title: {
+			type: DataTypes.STRING,
+			allowNull: false,
+		},
+		artist: {
+			type: DataTypes.STRING,
+			allowNull: false,
+		},
+		filename: {
+			type: DataTypes.STRING,
+			allowNull: false,
+		},
+		date_created: {
+			type: DataTypes.DATE,
+			allowNull: false,
+			defaultValue: DataTypes.NOW,
+		},
+		isPrivate: {
+			type: DataTypes.BOOLEAN,
       allowNull: false,
-      primaryKey: true,
-      autoIncrement: true,
-    },
-    title: {
-      type: DataTypes.STRING,
-      allowNull: false,
-    },
-    artist: {
-      type: DataTypes.STRING,
-      allowNull: false,
-    },
-    filename: {
-      type: DataTypes.STRING,
-      allowNull: false,
-    },
-     gallery_id: {
-      type: DataTypes.INTEGER,
-      references: {
-        model: 'usergallery',
-        key: 'id',
-      },
-    },
-  },
-  {
-    sequelize,
-    freezeTableName: true,
-    underscored: true,
-    modelName: 'image',
-  }
+      defaultValue: true,
+		},
+		gallery_id: {
+			type: DataTypes.INTEGER,
+			references: {
+				model: "usergallery",
+				key: "id",
+			},
+		},
+	},
+	{
+		sequelize,
+		freezeTableName: true,
+		underscored: true,
+		modelName: "image",
+	}
 );
 
 module.exports = Image;
